added dynamic download scripts

+++ ec2/ubuntu/ubuntu3/scripts/ubuntu.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/bash
 
 set -e
-set -x
 
 basic () {
 sudo apt-get update -y # && sudo apt-get upgrade -y
@@ -19,7 +18,7 @@ curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.1/install.sh | bash
 export NVM_DIR="$HOME/.nvm"
 [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
 [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
-# source ~/.bashrc
+source ~/.bashrc
 nvm install 14.7.0
 
 # sudo apt install golang -y
@@ -27,13 +26,13 @@ nvm install 14.7.0
 #sudo rm -rf /usr/local/go
 #curl -OL https://golang.org/dl/go1.16.7.linux-amd64.tar.gz
 curl -OL https://go.dev/dl/go1.18.5.linux-amd64.tar.gz
-sudo tar -C /usr/local -xvf go1.18.5.linux-amd64.tar.gz
+sudo tar -C /usr/local -xf go1.18.5.linux-amd64.tar.gz
 mkdir -p ~/go
 export GOPATH=$HOME/go
 export PATH=$PATH:/usr/local/go/bin:$GOPATH/bin
 echo 'export GOPATH=$HOME/go' >> ~/.bashrc
 echo 'export PATH=$PATH:/usr/local/go/bin:$GOPATH/bin' >> ~/.bashrc
-# source ~/.bashrc
+source ~/.bashrc
 go version
 }
 
@@ -46,18 +45,23 @@ docker () {
 #sudo apt-get update -y
 #sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-compose-plugin
 
-curl -fsSL https://get.docker.com -o /tmp/get-docker.sh
-sudo sh /tmp/get-docker.sh
-sudo usermod -a -G docker ubuntu
-sudo chmod 666 /var/run/docker.sock
-sudo systemctl enable docker
-sudo service docker start
-sudo systemctl status docker
-docker run hello-world
+if [[ ! -f $(which docker) ]]
+then
+  curl -fsSL https://get.docker.com -o /tmp/get-docker.sh
+  sudo sh /tmp/get-docker.sh
+  sudo usermod -a -G docker ubuntu
+  sudo chmod 666 /var/run/docker.sock
+  sudo systemctl enable docker --now
+  sudo systemctl status docker --no-pager
+  docker run hello-world
+fi
 
 
-sudo curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-sudo chmod +x /usr/local/bin/docker-compose
+if [[ ! -f $(which docker-compos) ]]
+then
+  sudo curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -s -o /usr/local/bin/docker-compose
+  sudo chmod +x /usr/local/bin/docker-compose
+fi
 docker-compose --version
 }
 
@@ -68,6 +72,17 @@ then
 git config --global user.name "Amit Karpe"
 git config --global user.email "amitkarpe@gmail.com"
 git config --global credential.username amitkarpe
+git config --global credential.host github.com
+git config --global credential.helper store
+git config --global credential.writelog true
+git config --global core.editor "vim"
+git config --global color.ui true
+git config --global color.ui auto
+git config --global color.status auto
+git config --global color.branch auto
+git config --global color.diff auto
+git config --global core.excludesfile ~/.gitignore_global
+
 export PAGER=''
 git config --global --list
 cat ~/.gitconfig
@@ -78,8 +93,8 @@ fi
 main () {
   basic
   dev
-  # docker
-  # git
+ docker
+ git
 }
 
 main
